lsirq,irqtop: cleanup threshold datatype

+++ sys-utils/irq-common.c
@@ -531,7 +531,7 @@ struct libscols_table *get_scols_table(struct irq_output *out,
 					      struct irq_stat *prev,
 					      struct irq_stat **xstat,
 					      int softirq,
-					      unsigned long threshold,
+					      uintmax_t threshold,
 					      size_t setsize,
 					      cpu_set_t *cpuset)
 {
@@ -570,7 +570,7 @@ struct libscols_table *get_scols_table(struct irq_output *out,
 	}
 
 	for (i = 0; i < stat->nr_irq; i++)
-		if (result[i].total >= threshold)
+		if ((uintmax_t) result[i].total >= threshold)
 			add_scols_line(out, &result[i], table);
 
 	free(result);
